+++ growthmate-api/src/main/java/com/growup/growthmate/GrowthMateApiApplication.java
@@ -1,4 +1,4 @@
-package com.growup;
+package com.growup.growthmate;
 
 import jakarta.annotation.PostConstruct;
 import org.springframework.boot.SpringApplication;
